requirements.txt
streamlit>=1.26
openai>=1.28.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
numpy>=1.24
